requirements.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aim==3.29.1
aim-ui==3.29.1
aimrecords==0.0.7
aimrocks==0.5.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aistudio_sdk==0.3.5
alembic==1.14.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
argcomplete==3.6.2
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrdict==2.0.1
attrs==25.3.0
babel==2.17.0
base58==2.0.1
bce-python-sdk==0.9.42
beautifulsoup4==4.13.4
blessed==1.20.0
blinker==1.8.2
boto3==1.37.38
botocore==1.37.38
cachetools==5.5.2
causal-conv1d @ file:///mnt/disk1/ruanwentao/code/MambaIR-v2/causal_conv1d-1.0.0%2Bcu118torch2.1cxx11abiTRUE-cp38-cp38-linux_x86_64.whl#sha256=1f70f62cb7cc974bea650ec78d65c13ba2a4b254b93d0879ee13e782fef138f8
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
clean-fid==0.1.35
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
contourpy==1.1.1
crcmod==1.7
cryptography==46.0.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.1.3
dataclasses==0.6
dataclasses-json==0.6.7
decorator==5.2.1
dependency-groups==1.3.1
diffuser==0.0.1
diffusers==0.32.2
distlib==0.3.9
distro==1.9.0
einops==0.8.1
entmax==1.3
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fairscale==0.4.13
fastapi==0.118.0
filelock==3.14.0
fire==0.7.1
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.2.3
future==1.0.0
fuzzywuzzy==0.18.0
fvcore==0.1.5.post20221221
google-auth==2.38.0
google-auth-oauthlib==1.0.0
gpustat==1.1.1
GPUtil==1.4.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.16.0
h5py==3.11.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
imageio==2.35.1
imagesize==1.4.1
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.1
jmespath==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.8
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
lanms_neo==1.0.2
lazy_loader==0.4
linformer==0.2.3
lmdb==1.6.2
lpips==0.1.4
lxml==6.0.0
Mako==1.3.10
mamba-ssm @ file:///mnt/disk1/ruanwentao/code/MambaIR-v2/mamba_ssm-1.0.1%2Bcu118torch2.1cxx11abiTRUE-cp38-cp38-linux_x86_64.whl#sha256=350be46927dac6ad415dd7f318dedd71ac32693d46572af06e96dc196accb841
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.5
mdurl==0.1.2
mmengine==0.10.7
model-index==0.1.11
modelscope==1.29.0
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
natsort==8.4.0
natten==0.17.1+torch240cu121
networkx==3.1
ninja==1.11.1.4
nox==2025.5.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.100.2
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.15
oss2==2.17.0
packaging==24.2
paddleocr==2.6.0.3
paddlepaddle==2.6.0
paddlex==3.2.0
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
Polygon3==3.0.9.1
pooch==1.8.2
portalocker==3.0.0
premailer==3.10.0
prettytable==3.11.0
propcache==0.2.0
protobuf==3.20.3
psutil==7.0.0
ptflops==0.7.3
ptwt==0.1.6
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
PyMuPDF==1.19.0
pyparsing==3.1.4
pypdfium2==4.30.0
pytesseract==0.3.13
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytorch-fid==0.3.0
pytorch-wavelets==1.3.0
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.2
rapidfuzz==3.9.7
rarfile==4.2
regex==2024.11.6
requests==2.28.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
RestrictedPython==7.4
rich==13.4.2
rsa==4.9
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.8
s3transfer==0.11.5
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
screen==1.0.1
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.44.0
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.7.10
tiktoken==0.7.0
timm==1.0.15
tokenizers==0.20.3
tomli==2.2.1
torch==2.4.1
torchaudio==2.4.1
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.19.1
tqdm==4.65.2
transformers==4.46.3
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2025.1
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.33.0
virtualenv==20.31.2
visualdl==2.5.3
warmup_scheduler==0.3
watchdog==4.0.2
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.6
yacs==0.1.8
yapf==0.43.0
yarl==1.15.2
zipp==3.20.2
